requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.9
flask-restplus==0.13.0
gunicorn==20.1.0
python-dotenv==0.19.2
python-pptx==0.6.21
requests==2.26.0
Werkzeug==0.16.0
